fix: structure of imports and typing
