+++ src/@types/components/index.ts
@@ -1,4 +1,4 @@
-import { Dispatch, SetStateAction } from 'react';
+import { Dispatch, ReactNode, SetStateAction } from 'react';
 
 export interface SettingsProps {
   handleChangeTheme: (value: string) => void;
@@ -7,5 +7,11 @@ export interface SettingsProps {
   handleHiddenUserPhoto: () => void;
   fontSize: number;
   setFontSize: Dispatch<SetStateAction<number>>;
+}
+
+export interface SocialMediaItemProps {
+  children: ReactNode;
+  text: string;
+  style?: any;
 }
