+++ src/@types/pages/index.ts
@@ -1,0 +1,9 @@
+export interface userDetailsProps {
+  avatar_url: string;
+  name: string;
+  login: string;
+  location: string;
+  bio: string;
+  twitter_username: string;
+  blog: string;
+}
