+++ src/components/SocialMediaItem/index.tsx
@@ -1,17 +1,17 @@
-import { ReactNode } from 'react'
-import * as S from './styles'
+import { SocialMediaItemProps } from '../../@types/components';
 
-export interface SocialMediaItemProps {
-  children: ReactNode
-  text: string
-  style?: any;
-}
+import * as S from './styles';
 
-export const SocialMediaItem = ({ children, text, style }: SocialMediaItemProps) => {
+export const SocialMediaItem = ({
+  children,
+  text,
+  style,
+}: SocialMediaItemProps) => {
   return (
     <S.Container>
       {children}
       <S.text style={style}>{text}</S.text>
     </S.Container>
-  )
-}
+  );
+};
+
